Revert "drop auto-adding the simplefb module to initramfs"
This reverts commit 9c41f9482666a392b80a3c4da3e695c4649d8ee1 as it
seems there are quite a few systems where only the simplefb is
available, so even though this sometimes causes issues it seems less
problematic if present than not.
And the issues are mostly restricted to passthrough and the host
grabbing the GPU through that module, with workarounds possible.

Update the filename while at it.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>